Fix configuration for jira_service_desk_derived.users_v1
