+++ sql/moz-fx-data-shared-prod/jira_service_desk/user/view.sql
@@ -1,0 +1,7 @@
+CREATE OR REPLACE VIEW
+  `moz-fx-data-shared-prod.jira_service_desk.user`
+AS
+SELECT
+  *
+FROM
+  `moz-fx-data-shared-prod.jira_service_desk_derived.user_v1`
