--- sql/moz-fx-data-shared-prod/jira_service_desk/users/metadata.yaml
@@ -1,14 +1,0 @@
-friendly_name: Jira Service Desk User
-description: |-
-   User data pulled from Jira API v3 (/rest/api/3/users/search)
-owners: [jmoscon@mozilla.com]
-labels:
-  schedule: daily
-  incremental: true
-  owner1: jmoscon@mozilla.com
-
-scheduling:
-  dag_name: bqetl_jira_user
-  arguments:
-    - --fivetran-jira-integration-jira-username={{ var.value.fivetran_jira_integration_jira_username }}
-    - --fivetran-jira-integration-jira-token={{ var.value.fivetran_jira_integration_jira_token }}
