+++ sql/moz-fx-data-shared-prod/jira_service_desk_derived/user_v1/metadata.yaml
@@ -1,0 +1,17 @@
+friendly_name: Jira Service Desk User
+description: |-
+   User data pulled from Jira API v3 (/rest/api/3/users/search)
+owners:
+- jmoscon@mozilla.com
+labels:
+  schedule: daily
+  incremental: false
+  owner1: jmoscon@mozilla.com
+
+scheduling:
+  dag_name: bqetl_jira_service_desk
+  secrets:
+  - deploy_target: JIRA_USERNAME
+    key: bqetl_jira_service_desk__jira_username
+  - deploy_target: JIRA_TOKEN
+    key: bqetl_jira_service_desk__jira_token
